+++ annotations/src/main/java/arez/annotations/ArezComponent.java
@@ -93,13 +93,18 @@ public @interface ArezComponent
   Feature observable() default Feature.AUTODETECT;
 
   /**
-   * Return true if the component should support implement {@link arez.component.DisposeTrackable}.
+   * Return enum to control whether the component should support implement {@link arez.component.DisposeTrackable}.
    * This will result in the component invoking dispose listener callbacks during dispose operation
    * within the scope of the disposing transaction.
    *
-   * @return true if the component should support dispose callbacks.
+   * <p>{@link Feature#ENABLE} will force the implementation of the DisposeTrackable interface,
+   * {@link Feature#DISABLE} will result in not implementing the DisposeTrackable interface and
+   * {@link Feature#AUTODETECT} will cause the component to implement the interface if the component
+   * is not annotated with <tt>javax.inject.Singleton</tt>.</p>
+   *
+   * @return Return enum to control whether the component should support implement DisposeTrackable.
    */
-  boolean disposeTrackable() default true;
+  Feature disposeTrackable() default Feature.AUTODETECT;
 
   /**
    * Return true if the component should dispose itself once it is no longer "observed".