+++ processor/src/test/resources/bad_input/com/example/dispose_trackable/NoDisposeTrackableWithRepositoryModel.java
@@ -2,10 +2,11 @@ package com.example.dispose_trackable;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
 import arez.annotations.Repository;
 
 @Repository
-@ArezComponent( disposeTrackable = false )
+@ArezComponent( disposeTrackable = Feature.DISABLE )
 public abstract class NoDisposeTrackableWithRepositoryModel
 {
   @Action