+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithDetachOnlyRepository.java
@@ -6,8 +6,6 @@ import arez.Component;
 import arez.Disposable;
 import arez.Observable;
 import arez.component.ComponentState;
-import arez.component.DisposeNotifier;
-import arez.component.DisposeTrackable;
 import arez.component.Identifiable;
 import java.util.stream.Stream;
 import javax.annotation.Generated;
@@ -19,7 +17,7 @@ import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
 @Singleton
-final class Arez_RepositoryWithDetachOnlyRepository extends RepositoryWithDetachOnlyRepository implements Disposable, Identifiable<Integer>, DisposeTrackable {
+final class Arez_RepositoryWithDetachOnlyRepository extends RepositoryWithDetachOnlyRepository implements Disposable, Identifiable<Integer> {
   private static volatile int $$arezi$$_nextId;
 
   private final int $$arezi$$_id;
@@ -30,8 +28,6 @@ final class Arez_RepositoryWithDetachOnlyRepository extends RepositoryWithDetach
   private final ArezContext $$arezi$$_context;
 
   private final Component $$arezi$$_component;
-
-  private final DisposeNotifier $$arezi$$_disposeNotifier;
 
   @Nonnull
   private final Observable<Stream<RepositoryWithDetachOnly>> $$arez$$_entities;
@@ -44,8 +40,7 @@ final class Arez_RepositoryWithDetachOnlyRepository extends RepositoryWithDetach
     if ( Arez.shouldCheckInvariants() ) {
       this.$$arezi$$_state = ComponentState.COMPONENT_INITIALIZED;
     }
-    this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().createComponent( "RepositoryWithDetachOnlyRepository", $$arezi$$_id(), Arez.areNamesEnabled() ? $$arezi$$_name() : null, () -> $$arezi$$_preDispose() ) : null;
-    this.$$arezi$$_disposeNotifier = new DisposeNotifier();
+    this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().createComponent( "RepositoryWithDetachOnlyRepository", $$arezi$$_id(), Arez.areNamesEnabled() ? $$arezi$$_name() : null, () -> super.preDispose() ) : null;
     this.$$arez$$_entities = $$arezi$$_context().createObservable( Arez.areNativeComponentsEnabled() ? this.$$arezi$$_component : null, Arez.areNamesEnabled() ? $$arezi$$_name() + ".entities" : null, Arez.arePropertyIntrospectorsEnabled() ? () -> super.entities() : null, null );
     if ( Arez.shouldCheckInvariants() ) {
       this.$$arezi$$_state = ComponentState.COMPONENT_CONSTRUCTED;
@@ -85,17 +80,6 @@ final class Arez_RepositoryWithDetachOnlyRepository extends RepositoryWithDetach
     return "RepositoryWithDetachOnlyRepository";
   }
 
-  private void $$arezi$$_preDispose() {
-    super.preDispose();
-    $$arezi$$_disposeNotifier.dispose();
-  }
-
-  @Override
-  @Nonnull
-  public DisposeNotifier getNotifier() {
-    return $$arezi$$_disposeNotifier;
-  }
-
   @Override
   public boolean isDisposed() {
     return ComponentState.isDisposingOrDisposed( this.$$arezi$$_state );
@@ -109,7 +93,7 @@ final class Arez_RepositoryWithDetachOnlyRepository extends RepositoryWithDetach
         this.$$arezi$$_component.dispose();
       } else {
         $$arezi$$_context().safeAction( Arez.areNamesEnabled() ? $$arezi$$_name() + ".dispose" : null, () -> { {
-          this.$$arezi$$_preDispose();
+          super.preDispose();
           this.$$arez$$_entities.dispose();
         } } );
       }