+++ processor/src/test/resources/input/com/example/dispose_trackable/NoDisposeTrackableModel.java
@@ -2,8 +2,9 @@ package com.example.dispose_trackable;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
 
-@ArezComponent( disposeTrackable = false )
+@ArezComponent( disposeTrackable = Feature.DISABLE )
 public abstract class NoDisposeTrackableModel
 {
   @Action